requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==20.0.4
whitenoise==6.7.0
setuptools==80.0.0
